Fixed date picker timezone issues
The update to the date picker for multiple days broke the previous/next buttons in negative time zones again.  Also, I discovered an old issue with the previous button causing a one month backward jump from the current day if it was the first day of the month.  Also, there was in issue with the timeline data being one day behind the date picker in negative time zones.